+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Enhancements
 
+* Set the default strategy for partitioning images to `hi_res`
+
 ### Features
 
 ### Fixes